+++ packages/shared/src/index.ts
@@ -4,6 +4,7 @@ export * from './types/post.types';
 export * from './types/course.types';
 export * from './types/event.types';
 export * from './types/theme.types';
+export * from './types/notification.types';
 
 // Utils
 export * from './utils/date.utils';